requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
promise==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
twisted==23.10.0
txaio==23.1.1
typing-extensions==4.8.0
urllib3==2.0.6
zope.interface==6.1
